refactor : carsService 클래스 삭제 및 carsService의 메서드 Cars로 이동
