+++ src/main/java/racingcar/model/Cars.java
@@ -1,6 +1,8 @@
 package racingcar.model;
 
 import java.util.ArrayList;
+import java.util.Collections;
+import java.util.LinkedHashMap;
 import java.util.List;
 import racingcar.constant.Error;
 import racingcar.util.Utils;
@@ -14,12 +16,38 @@ public class Cars {
         carList = generateCarList(carNames);
     }
 
-    public Car getCar(int index) {
-        return carList.get(index);
-    }
-
     public int size() {
         return carList.size();
+    }
+
+    public void moveCar(int index) {
+        carList.get(index).addPosition();
+    }
+
+    public int getMaxPosition() {
+        List<Integer> carPositionList = new ArrayList<>();
+        for (Car car : carList) {
+            carPositionList.add(car.getPosition());
+        }
+        return Collections.max(carPositionList);
+    }
+
+    public List<String> getCarNamesAtPosition(int position) {
+        List<String> carNamesAtPosition = new ArrayList<>();
+        for (Car car : carList) {
+            if (car.getPosition() == position) {
+                carNamesAtPosition.add(car.getName());
+            }
+        }
+        return carNamesAtPosition;
+    }
+
+    public LinkedHashMap<String, Integer> getCarNamePositionMap() {
+        LinkedHashMap<String, Integer> carNamePositionMap = new LinkedHashMap<>();
+        for (Car car : carList) {
+            carNamePositionMap.put(car.getName(), car.getPosition());
+        }
+        return carNamePositionMap;
     }
 
     private List<Car> generateCarList(String[] carNames) {
